requirements.txt
# NovaOS dashboard + memory
streamlit>=1.36
pandas>=2.2
altair>=5.1
plotly>=5.22
redis>=5.0
python-dotenv>=1.0
psycopg[binary]>=3.1

# Existing NovaOS stack
langchain-openai==0.1.25
langgraph==0.2.34
langchain-community==0.2.17
dropbox==12.0.2
openai==1.51.2
fastapi==0.116.1
uvicorn==0.35.0
psycopg[binary]==3.2.9
